Move validators into correct folder
